game-flow: add story so far to TR2X
